Add getBoardList
