+++ src/apis/board/index.ts
@@ -1,2 +1,3 @@
 export * from './comment';
 export * from './getBoardDetail';
+export * from './getBoardList';
