+++ src/app/community/board/page.tsx
@@ -1,9 +1,5 @@
-import { cookies } from 'next/headers';
-import { redirect } from 'next/navigation';
-
-import { boardUrl } from '@/libs';
+import { getBoardList } from '@/apis';
 import { Board } from '@/pageContainer';
-import type { BoardInfoType } from '@/types';
 
 import type { Metadata } from 'next';
 
@@ -20,32 +16,6 @@ const BoardPage = async () => {
   const boardList = await getBoardList();
 
   return <Board initialData={[...boardList]} />;
-};
-
-const getBoardList = async (): Promise<BoardInfoType[]> => {
-  const accessToken = cookies().get('accessToken')?.value;
-
-  if (!accessToken) return redirect('/auth/refresh');
-
-  const response = await fetch(
-    new URL(`/api/v1${boardUrl.getBoardList(0)}`, process.env.BASE_URL),
-    {
-      method: 'GET',
-      headers: { Cookie: `accessToken=${accessToken}` },
-    }
-  );
-
-  if (response.status === 401) {
-    return redirect('/auth/refresh');
-  }
-
-  if (!response.ok) {
-    return redirect('/auth/signin');
-  }
-
-  const boardList: BoardInfoType[] = await response.json();
-
-  return boardList;
 };
 
 export default BoardPage;
